Add sparkml for machine learning class
Similar to centos7-ml, this image contains all we need to run HappyCat.
The main difference is, the base image is changed to alpine linux, which
is much smaller (only ~300MB) and faster to launch.

Build & install image:

  $ docker build -t sparkml - < sparkml.dockerfile

Image usage (mount scripts and sample data to container's /ml and /ml_data):

  $ git clone https://github.com/sparktsao/HappyCat.git
  $ docker run -v HappyCat/learning:/ml -v HappyCat/dataset/unittest:/ml_data \
               --name happycat --rm -t -i sparkml python happycat.py /ml_data/ 2 unittest

Please refer to the comments in dockerfile for customization.